// core/api/Platform.ts
/* tslint:disable:no-unused-vars */
/* eslint-disable no-unused-vars */
enum Platform {
  /* tslint:disable:no-unused-vars */
  UBER = 'UBER',
  /* tslint:disable:no-unused-vars */
  TAXI = 'TAXI',
  /* tslint:disable:no-unused-vars */
  INDRIVE = 'INDRIVE',
  /* tslint:disable:no-unused-vars */
  DIDI = 'DIDI',
  /* tslint:disable:no-unused-vars */
  BEAT = 'BEAT',
  /* tslint:disable:no-unused-vars */
  CABIFY = 'CABIFY',
  /* tslint:disable:no-unused-vars */
  PERSONAL = 'PERSONAL',
  /* tslint:disable:no-unused-vars */
  OTHER = 'OTHER',
}
/* tslint:disable:no-unused-vars */
export { Platform };